Add files via upload
Harnessing the power of Convolutional Neural Networks (CNN) and Python libraries like TensorFlow and Keras, this app transforms images of vegetables into a delectable array of 10 tailored recipes. Crafted at the crossroads of my culinary past and data expertise, it's a fusion of flavors and analytics. Dive into the code on GitHub and explore the magic that unfolds when gastronomy meets technology!